requirements.txt
torchvision==0.15.2+cu118
tqdm==4.56.0
transformers==4.41.1
triton==2.0.0
typer==0.4.2
types-pytz==2024.1.0.20240417
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wasabi==0.10.1